prevent use of rspec-puppet 1.0.0 or later
Version 1.0.0 of rspec-puppet apparently made some changes to how
catalogues can be queried (among other things), which has broken a
number of tests that made use of some undocumented and unsupported APIs.
There may be a workaround but I'm not well-versed enough to apply it, so
instead I've changed the Gemfile to prevent new versions of this
library from being used.

This is the only different between branch 3.2.1 and the repo at tag
3.2.0.